backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
nfcpy==1.0.4